+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - feature: new supported `MC 1.20` version!
 - [chat.history] fix: make it working in `offline-mode` server.
 - [command_permission] feature: now we can detect `in-complete command` or `in-correct argument` when using `/command-permission describe <cmd>` command.
+- [system_message] feature: add debug logging when replacing a translatable text.